:green_heart: Fix CI error where running the migration command raises an exception.
The exception is raised because the migrate command was performing system checks,
which causes to check for any system problems. Since during the definition of one
of the views we were attempting to get some data from the settings table which
does not yet exist using a utility class, an exception would be raised. To circumvent
this problem, we just need to let the utility catch the exception we get and just
give back the default value specified by the view.